app/pyproject.toml
[project]
name = "gcs-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp",
    "google-cloud-storage",
    "google-api-core",
    "pydantic",
    "uvicorn",
    "packaging"
]